Added integration test
